Add main file, hash file, travis, webpack, package.json, .gitignore and spec files.
Basically, this is the initial commit.
The main file extends Gun with the chain and constructor methods, allowing you to set a consistent prefix for all instances, and tweak each instance individually.
The hash file takes a string and generates a deterministic, low-collision prefix from the input. This prevents someone from accidentally colliding with the namespace by using the same name in their key as the scope.