Document use in Node and the browser, bump version for npm
The documentation mentioned nothing about how Reticle is used in node, and the behavior is noticably different. Revisions have been made to clear that up.
Reticle was never published to npm after the initial draft was finished, so `.key` still has no support. The version number has been bumped in preparation for publishing.