Fix numerous crippling bugs, simplify code, overload default UUID generator, improve nodejs interface, add tests.
Any pointers contained within an object wouldn't be properly scoped, meaning that keys couldn't contain objects (I wasn't kidding when I said it was crippling!). This was partly because I hadn't accounted for the UUIDs not being scoped as well, so when the request hit the overloaded `.get` method, it was prefixed and the original pointer was lost. I've since added a scoping mechanism to the UUID generator that scopes graphs according to the chain context, fixing the pointer referencing and nesting issues. Yay!

Some code for looping was unnecessarily complex, and so I replaced it with an smaller, easier to digest function.

Due to my obliviousness, I missed the bit where node will recursively travel up the folder tree until it finds the package it's looking for. I'd created a redundant/verbose system that node already abstracted away for me. Now instead of exporting a function, it will export the Gun constructor (if it's installed in your project, and if it isn't, what good is reticle?).

More tests to pass! And all of them are!! Yay!!!